Steganography-master/requirements.txt
pillow>=3.1.1
